added package json and split matrix code
