+++ index.html
@@ -211,7 +211,8 @@
 	<script type="text/javascript" src="jquery.csv.min.js"></script>
 	<script type="text/javascript" charset="utf8" src="https://cdn.datatables.net/1.10.22/js/jquery.dataTables.js">
 	</script>
-	<script type='text/javascript' src='main.js'></script>
+	<script type='module' src='matrix.js'></script>
+	<script type='module' src='main.js'></script>
 
 </body>
