+++ package.json
@@ -1,0 +1,18 @@
+{
+    "name": "network_navigator",
+    "description": "",
+    "version": "2.0.0",
+    "main": "main.js",
+    "scripts": {
+      "start": "npx http-server"
+    },
+    "repository": {
+      "type": "git",
+      "url": "https://github.com/jrladd/network_navigator.git"
+    },
+    "keywords": [],
+    "author": "",
+    "license": "ISC",
+    "homepage": "https://network-navigator.library.cmu.edu/",
+    "type": "module"
+  }
